Got the physics stuff sort of working.
Sort of working, anyway. It's not as springy as intended. But it mostly works. I think it breaks if it rotates in more than one axis at a time, so that's fun. I'll probably continue to work on this, or at this point, I might just give up and use the built-in Unity RigidBodies.